No result view is used on company screen in an appropriate case
