+++ mobile/src/navigation/loading.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ActivityIndicator, View } from 'react-native';
-import { searchViewStyles as styles } from './search-view-styles';
+import { searchViewStyles as styles } from './search/search-view-styles';
 
 //============================================================================
 export class LoadingView extends React.Component {
