+++ mobile/src/people/employees-list.tsx
@@ -3,17 +3,21 @@ import { ListRenderItemInfo, View } from 'react-native';
 
 import { Employee } from '../reducers/organization/employee.model';
 import { EmployeesListItem } from './employees-list-item';
-import { employeesListStyles as styles, noResult } from './styles';
+import { employeesListStyles as styles } from './styles';
 import { employeesAZComparer } from './employee-comparer';
 import { FlatList } from 'react-navigation';
-import { StyledText } from '../override/styled-text';
+import { NoResultView } from '../navigation/search/no-result-view';
 
+//============================================================================
 export interface EmployeesListProps {
     employees: Employee[];
     onItemClicked: (e: Employee) => void;
 }
 
+//============================================================================
 export class EmployeesList extends React.Component<EmployeesListProps> {
+
+    //----------------------------------------------------------------------------
     public render() {
         const employees = this.props.employees.sort(employeesAZComparer);
 
@@ -22,21 +26,15 @@ export class EmployeesList extends React.Component<EmployeesListProps> {
                 data={employees}
                 keyExtractor={this.keyExtractor}
                 renderItem={this.renderItem}
-                ListEmptyComponent={this.listEmptyComponent()}
+                ListEmptyComponent={<NoResultView/>}
             />
         </View>;
     }
 
-    private listEmptyComponent() {
-        return (
-            <View style={noResult.container}>
-                <StyledText style={noResult.text}>{'No result'}</StyledText>
-            </View>
-        );
-    }
-
+    //----------------------------------------------------------------------------
     private keyExtractor = (item: Employee) => item.employeeId;
 
+    //----------------------------------------------------------------------------
     private renderItem = (itemInfo: ListRenderItemInfo<Employee>) => {
         const { item } = itemInfo;
