+++ mobile/src/people/people-screen.tsx
@@ -3,7 +3,7 @@ import { createStackNavigator, NavigationRouteConfigMap } from 'react-navigation
 import { EmployeeDetailsScreen } from '../employee-details/employee-details-screen';
 import { stackNavigatorConfig } from '../override/stack-navigator-config';
 import { PeopleScreenNavigator } from './navigator/people-screen-navigator';
-import { SearchViewPeople } from '../navigation/search-view';
+import { SearchViewPeople } from '../navigation/search/search-view';
 import { CurrentPeopleDepartment } from './current-people-department';
 import { CurrentPeopleRoom } from './current-people-room';
